v1.4 'Ariadnes Threads'
Added Threading to Mod Initializing boosting Performance
Added Expander to Performance Watchers to combat Window being too large
Added Thread Watcher to Performance Watchers
Added better system logging to startup
Moved the UnhandledException Handler to the top to prevent deleting the old log file from causing the App to crash unexpectedly
Added MessageBox to show when an app crashed with the option to clear the cache
Added RAR support via unrar.dll: It includes both 32Bit and 64Bit version, which is a bit redudant, but will load the necessary file on startup by determining both System and Program version
Cleaned up Logger and added support for threaded logging
Made ModNames sorted in the UI, taking a bit of a performance hit though
Added options menu
Added deleteOldFiles and copyAllFiles options to the options menu and the User-Settings
Cleaned up the MainWindow and its functions
Streamlined the process of finding which files to merge
Moved Mod-related files to subdirectory "Mods"
Made ModManager create an index of all vanilla files instead of opening and searching for every single one. Increases performance by a significant margin
Improved indexing by doing pak->files rather than file->pak
Streamlined Manifest reading
Fixed bug causing empty tags in the manifest to crash the app
Added ModManifest class which manages the manifest alone, making it easier to add more tags in the future
Added UIUtilities for common functions (and maybe renaming it later)
Made buttons that do not have the requirements only disabled and not hidden, displaying requirements as a tooltip
Improved performance tenfolds by explicitly calling the Garbage Collector after a merge operation. There's  lot of stuff that will be created during merge, due to the nature of it, which will cause the GC to only work after a set amount of time or even later depending on certain conditions. By calling it explicitly, it will clean up all the memory that will never be used anyways
Added Startup lua script to merged mods mod which will display that the mod is loaded and the files it contains on startup in the console (stole that line from cheat mod). It shouldn't conflict with anything, since it uses its own name as the file name of the lua file.